+++ src/signalling/elements/endpoints/webrtc/publish_endpoint.rs
@@ -11,7 +11,7 @@ use medea_control_api_proto::grpc::api as proto;
 
 use crate::{
     api::control::endpoints::webrtc_publish_endpoint::{
-        P2pMode, WebRtcPublishId as Id,
+        AudioSettings, P2pMode, VideoSettings, WebRtcPublishId as Id,
     },
     signalling::elements::{
         endpoints::webrtc::play_endpoint::WeakWebRtcPlayEndpoint,
@@ -38,6 +38,12 @@ struct WebRtcPublishEndpointInner {
 
     /// Owner [`Member`] of this [`WebRtcPublishEndpoint`].
     owner: WeakMember,
+
+    /// Settings for the audio media type of the [`WebRtcPublishEndpoint`].
+    audio_settings: AudioSettings,
+
+    /// Settings for the video media type of the [`WebRtcPublishEndpoint`].
+    video_settings: VideoSettings,
 
     /// [`PeerId`] of all [`Peer`]s created for this [`WebRtcPublishEndpoint`].
     ///
@@ -115,6 +121,8 @@ impl WebRtcPublishEndpoint {
         p2p: P2pMode,
         owner: WeakMember,
         is_force_relayed: bool,
+        audio_settings: AudioSettings,
+        video_settings: VideoSettings,
     ) -> Self {
         Self(Rc::new(RefCell::new(WebRtcPublishEndpointInner {
             id,
@@ -122,6 +130,8 @@ impl WebRtcPublishEndpoint {
             is_force_relayed,
             sinks: Vec::new(),
             owner,
+            audio_settings,
+            video_settings,
             peer_ids: HashSet::new(),
         })))
     }
@@ -203,6 +213,16 @@ impl WebRtcPublishEndpoint {
         self.0.borrow().is_force_relayed
     }
 
+    /// Returns [`AudioSettings`] of this [`WebRtcPublishEndpoint`].
+    pub fn audio_settings(&self) -> AudioSettings {
+        self.0.borrow().audio_settings
+    }
+
+    /// Returns [`VideoSettings`] of this [`WebRtcPublishEndpoint`].
+    pub fn video_settings(&self) -> VideoSettings {
+        self.0.borrow().video_settings
+    }
+
     /// Downgrades [`WebRtcPublishEndpoint`] to weak pointer
     /// [`WeakWebRtcPublishEndpoint`].
     pub fn downgrade(&self) -> WeakWebRtcPublishEndpoint {
@@ -246,6 +266,8 @@ impl Into<proto::WebRtcPublishEndpoint> for WebRtcPublishEndpoint {
             p2p: p2p as i32,
             id: self.id().to_string(),
             force_relay: self.is_force_relayed(),
+            audio_settings: Some(self.audio_settings().into()),
+            video_settings: Some(self.video_settings().into()),
             on_stop: String::new(),
             on_start: String::new(),
         }