+++ jason/CHANGELOG.md
@@ -44,6 +44,7 @@ All user visible changes to this project will be documented in this file. This p
         - Ability to configure local media stream used by `Room` via `Room.set_local_media_settings()` ([#54], [#97]);
         - `Room.on_failed_local_stream` callback ([#54]);
         - `Room.on_close` callback for WebSocket close initiated by server ([#55]);
+    - Optional tracks support ([#106]);
     - `RtcIceTransportPolicy` configuration ([#79]).
 - Room management:
     - Library API:
@@ -79,6 +80,7 @@ All user visible changes to this project will be documented in this file. This p
 [#87]: /../../pull/87
 [#90]: /../../pull/90
 [#97]: /../../pull/97
+[#106]: /../../pull/106